Revert "Update v3.5.6928.357"
This reverts commit 3bb2eb261360fc36ae9de98d3f0ecd44985bdc4b.

+++ LeagueFPSBoostGUI/Configuration/AppConfig.cs
@@ -92,12 +92,11 @@ namespace LeagueFPSBoost.Configuration
                 Settings.Default.Reload();
             }
 
-            appConfigLogger.Warn("Deleting temporary configuration file has been disabled due to repeated windows installer repair.");
-            /*if (File.Exists(configFile))
+            if (File.Exists(configFile))
             {
                 appConfigLogger.Debug("Deleting temporary configuration file: " + configFile);
                 File.Delete(configFile);
-            }*/
+            }
         }
 
         /// <summary>
